Don't store ratings in the database
Previously, we used to dynamically calculate the rating of a
service when fetching that service's data, and then insert that
rating into the database. This resulted in an insert query being
ran for every service listed on the homepage (i.e. every service)
when visiting the homepage.

A potential solution could be to recalculate the rating for the
appropriate service whenever a point is changed. This would mean
the service ratings would not have to be recalculated as often,
and you could theoretically fetch a service and its rating without
having to fetch all related points. However, since I'm not
well-versed enough in Ruby/Ruby on Rails yet, I don't quite know
how to do this by updating the Points model. Furthermore, I'm not
sure what repercussions it would have elsewhere in the codebase.

Since we were _already_ recalculating the grades whenever they
were needed, I've simply removed the `grade` column from the
database, and removed the logic that re-inserted it into the
database after every recalculation. This should solve at least
our most egregious performance issues on the homepage.