Remove some magic values, hover effect
The card headings on the homepage showed a scrollbar on Firefox.
To prevent this, I removed the fixed height, set overflow to
hidden (with an ellipsis), and added the service name to the
`title` attribute so it is still fully visible on hover (on
devices that support hovering).

I also removed the hover effects over cards, because it looked
really weird with those changes and implied that you could
interact with the cards even though you cannot.