Save selections back to the database
The previous method of checking which text was selected
(.selectionStart and .selectionEnd) only worked on textareas,
which we removed. This now works again.

Also, the HTML of the scraped website was displayed - this commit
makes it so it's displayed as text.

The thing that doesn't work yet, and for which I think there's no
API available, it _setting_ the text selection to whatever was
previously selected.