Remove ratings from points
Ratings are now deduced from the case linked to a point. I'm
hoping that this does not introduce too many performance issues due
to JOINing cases everywhere we need a point's rating (which I think
is pretty much everywhere), and that I managed to fix all the
places where a point's rating was referred to.
This also means that points without a case will no longer have a
rating. Be sure to backup the database before applying this, since
it will remove the rating column from the points table, including
all its contents.

Fixes #503, #504.