Add PhantomJS to the Docker image
PhantomJS is used by the document crawler, which expects it to be
available on the path.

Note that we had to compile it from source. Although phantomjs is
available in the jessie-backports repository (which is why I
explicitly configured the image to be a Jessie image - it's
relevant to know which version of Debian we're running when adding
repositories), it doesn't work in Docker:
https://github.com/ariya/phantomjs/issues/14376#issuecomment-239687524