Only link to the annotation view with documents
In the service list, only show a link to the annotation view when
documents have already been defined for that service.
This has two advantages:

- Since adding documents currently requires knowledge of XPaths,
  this makes it easy for contributors to see which services they
   can immediately start annotating.
- For people who do understand XPath, it's easier to see which
  services should still have XPaths added to them.

Longer-term, when annotation is the default mode, we can probably
show them all the time again. In fact, it'd probably be the
default target when you click the service name.

Fixes #623.