Literally display HTML entities in a document
Although it's a shame that they are shown as such to the user, I
don't think there's a way of showing the referenced character while
having the HTML entity code be included in the selection (like
HTML tags are hidden but still included in the selection). Thus,
simply display the original HTML entity code.

Fixes #684.